functions/requirements.txt
firebase-admin==6.6.0
firebase-functions
fastapi
aiohttp
python-multipart
uvicorn
mangum
python-dotenv
